moving to shared styles, stub/modify data to help with consistency, @todo: clean-up/gather data
