+++ src/components/Feedel.tsx
@@ -2,12 +2,14 @@ import React from 'react';
 import { TouchableOpacity, Text, Animated, ImageStyle, TextStyle, ViewStyle } from 'react-native';
 
 interface FeedelProps {
-    item: Feedel;
+    item: {
+        title: string;
+    }
     onPress: () => void;
     styles: {
         item: ViewStyle;
         itemImage: ImageStyle;
-        title: TextStyle;
+        letter: TextStyle;
     };
 }
